Added and implemented ItemType + huge code cleanup. - Ordered implemented methods by their interface and declaration order. - Some whitespace cleanup. - Return double values in methods that should return doubles. - Changed abstract classes to final classes with private constructors.
  Making a class abstract does not prevent sub-classing and is like
  saying: "Please extend me."
  Making a class final with a private constructor prevents both
  instantiating and sub-classing.

+++ src/main/java/me/KiwiLetsPlay/KiwiField/game/StatsUtil.java
@@ -114,7 +114,7 @@ public class StatsUtil {
 		int kills = getKills(p);
 		int deaths = getDeaths(p);
 		if (deaths == 0) deaths = 1;
-		return kills / deaths;
+		return ((double) kills) / ((double) deaths);
 	}
 	
 	public int getMoney(Player player) {
